scheduler/requirements/production.txt
Django==2.1.11
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.2.0
django-allauth==0.38.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-uuidfield==0.5.0

oauthlib==3.0.1
PyJWT==1.7.1
pyfcm== 1.4.7

mysqlclient==1.3.13
python-dateutil==2.8.0
uWSGI==2.0.18
python3-openid==3.1.0

urllib3>=1.24.2

requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
